pyproject.toml
[tool.poetry]
name = "bmo"
version = "0.2.0"
description = "BMO automates stuff at SubCom"
authors = ["Dilawar Singh <dilawar@subcom.tech>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.8"
speedtest-cli = "^2.1.3"
typer = "^0.4.1"
msgpack = "^1.0.4"
requests = {extras = ["security"], version = "^2.28.0"}
sentry-sdk = "^1.6.0"

[tool.poetry.dev-dependencies]
mypy = "^0.960"
black = "^22.3.0"
pylint = "^2.14.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bmo = "bmo.main:app"
